Varnam editor component
