+++ Editor/package.json
@@ -14,6 +14,7 @@
     "postcss": "^8.4.38",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.1.0",
     "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
