+++ Editor/src/App.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 #root {
-  max-width: 1280px;
+  
   /* margin: 0 auto; */
   padding: 2rem;
   /* text-align: center; */
